refactor: 댓글 섹션 분리 및 셸(Shell) 패턴 적용
